requirements.txt
asgiref==3.6.0
certifi==2021.10.8
dbbackup==0.1.13
dbbackup-tools==0.1.5
Django==4.1.6
django-admin-honeypot==1.1.0
django-crontab==0.7.1
django-userforeignkey==0.5.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.6
python-decouple==3.6
pytz==2022.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.7
urllib3==1.26.9
